refactor: extract file and folder CRUD handlers to lib/fileOps.ts
Move create/rename/delete/duplicate operations for .http files and
folders out of App.svelte into src/lib/fileOps.ts, built on the typed
tree helpers in lib/tree.ts. TreeSidebar event handlers in App.svelte
become thin arrows that unwrap CustomEvent details. No behavior
change.
